+++ src/main/java/com/fourseason/delivery/domain/member/entity/Member.java
@@ -1,6 +1,6 @@
 package com.fourseason.delivery.domain.member.entity;
 
-import com.fourseason.delivery.global.auth.dto.SignUpRequestDto;
+import com.fourseason.delivery.global.auth.dto.request.SignUpRequestDto;
 import com.fourseason.delivery.global.entity.BaseTimeEntity;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
