+++ src/main/java/com/fourseason/delivery/global/auth/service/AuthService.java
@@ -3,8 +3,8 @@ package com.fourseason.delivery.global.auth.service;
 import com.fourseason.delivery.domain.member.entity.Member;
 import com.fourseason.delivery.domain.member.repository.MemberRepository;
 import com.fourseason.delivery.global.auth.JwtUtil;
-import com.fourseason.delivery.global.auth.dto.SignInRequestDto;
-import com.fourseason.delivery.global.auth.dto.SignUpRequestDto;
+import com.fourseason.delivery.global.auth.dto.request.SignInRequestDto;
+import com.fourseason.delivery.global.auth.dto.request.SignUpRequestDto;
 import com.fourseason.delivery.global.auth.dto.TokenDto;
 import com.fourseason.delivery.global.exception.CustomException;
 import lombok.RequiredArgsConstructor;
